fix(tui): include ocean palette in cache identity
Replace the randomized fingerprint that omitted ramp colors with a typed identity and explicit deterministic color encoding. This prevents visually different columns from reusing stale cached rows while preserving stable keys for identical semantic inputs.

Move the existing ocean tests into the structure-ratchet-excluded child test module and add coverage for every ramp color.
